+++ src/_tests_/init.spec.lua
@@ -17,21 +17,16 @@ return function()
 
 	it("should return an appropriate table for Context", function()
 		expect(type(RoactNavigation.Context)).to.equal("table")
-		expect(type(RoactNavigation.Context.Provider)).to.equal("function")
-		expect(type(RoactNavigation.Context.Consumer)).to.equal("function")
-		expect(type(RoactNavigation.Context.connect)).to.equal("function")
+		expect(type(RoactNavigation.Context.Provider)).to.equal("table")
+		expect(type(RoactNavigation.Context.Consumer)).to.equal("table")
 	end)
 
-	it("should return a function for Provider", function()
-		expect(type(RoactNavigation.Provider)).to.equal("function")
+	it("should return a Component for Provider", function()
+		expect(type(RoactNavigation.Provider)).to.equal("table")
 	end)
 
-	it("should return a function for Consumer", function()
-		expect(type(RoactNavigation.Consumer)).to.equal("function")
-	end)
-
-	it("should return a function for connect", function()
-		expect(type(RoactNavigation.connect)).to.equal("function")
+	it("should return a Component for Consumer", function()
+		expect(type(RoactNavigation.Consumer)).to.equal("table")
 	end)
 
 	it("should return a function for withNavigation", function()
@@ -87,7 +82,6 @@ return function()
 	end)
 
 	it("should return a valid component for EventsAdapter", function()
-		expect(RoactNavigation.EventsAdapter.render).never.to.equal(nil)
 		local instance = Roact.mount(Roact.createElement(RoactNavigation.EventsAdapter, {
 			navigation = {
 				addListener = function()